requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Reuploaded==1.4.0
PyPDF2==3.0.1
python-docx==1.1.0
pdfkit==1.0.0
schedule==1.2.0
